Moved mock async payment method
